refactoring into slots & useDrawer & stlying
