+++ packages/experimental/DrawerV1/src/DrawerV1.tsx
@@ -1,12 +1,9 @@
 /** @jsx withSlots */
-import { useState } from 'react';
-import { Button, ScrollView, StyleSheet, View } from 'react-native';
+import { Animated, Modal, TouchableWithoutFeedback, View, StyleSheet } from 'react-native';
 
 import type { UseSlots } from '@fluentui-react-native/framework';
 import { compose, mergeProps, withSlots } from '@fluentui-react-native/framework';
-import { Text } from '@fluentui-react-native/text';
 
-import Drawer from './Drawer';
 import { stylingSettings } from './DrawerV1.styling';
 import type { DrawerV1Type, DrawerV1Props } from './DrawerV1.types';
 import { DrawerV1Name } from './DrawerV1.types';
@@ -16,42 +13,41 @@ export const DrawerV1 = compose<DrawerV1Type>({
   displayName: DrawerV1Name,
   ...stylingSettings,
   slots: {
-    root: ScrollView,
+    root: View,
+    modal: Modal,
+    backdrop: TouchableWithoutFeedback,
+    backdropContent: Animated.View,
+    content: Animated.View,
+    dragger: View,
   },
   useRender: (userProps: DrawerV1Props, useSlots: UseSlots<DrawerV1Type>) => {
     const DrawerV1 = useDrawerV1(userProps);
     const Slots = useSlots(userProps);
 
-    const [isDrawerVisible, setIsDrawerVisible] = useState(false);
-    const [drawerPosition, setDrawerPosition] = useState('left');
-
-    const handleDrawerClose = () => {
-      setIsDrawerVisible(false);
-    };
-
-    const handleOpenDrawer = (position) => {
-      setDrawerPosition(position);
-      setIsDrawerVisible(true);
-    };
-    return (final: DrawerV1Props) => {
-      const { ...mergedProps } = mergeProps(DrawerV1.props, final);
-
+    return (final: DrawerV1Props, children: React.ReactNode) => {
+      const { isVisible, handleClose, handleBackdropPress, animatedOpacity, animatedStyle, position, ...rest } = mergeProps(
+        DrawerV1.props,
+        final,
+      );
       return (
-        <Slots.root {...mergedProps}>
-          <View style={styles.container}>
-            <View style={styles.content}>
-              <Text style={styles.text}>This is the main content of the screen.</Text>
-              <Button title="Open Drawer (Left)" onPress={() => handleOpenDrawer('left')} />
-              <Button title="Open Drawer (Right)" onPress={() => handleOpenDrawer('right')} />
-              <Button title="Open Drawer (Bottom)" onPress={() => handleOpenDrawer('bottom')} />
-            </View>
-            <Drawer isVisible={isDrawerVisible} onClose={handleDrawerClose} position={drawerPosition}>
-              <View style={styles.drawerContent}>
-                <Text style={styles.text}>This is the content of the drawer.</Text>
-                <Button title="Close Drawer" onPress={handleDrawerClose} />
-              </View>
-            </Drawer>
-          </View>
+        <Slots.root {...rest}>
+          {isVisible && (
+            <Slots.modal
+              visible={isVisible}
+              onRequestClose={handleClose}
+              supportedOrientations={['portrait', 'portrait-upside-down', 'landscape', 'landscape-left', 'landscape-right']}
+              animationType="none"
+              transparent
+            >
+              <Slots.backdrop onPress={handleBackdropPress}>
+                <Slots.backdropContent style={[{ opacity: animatedOpacity }]} />
+              </Slots.backdrop>
+              <Slots.content style={[styles[position], animatedStyle]}>
+                {position === 'bottom' && <Slots.dragger />}
+                {children}
+              </Slots.content>
+            </Slots.modal>
+          )}
         </Slots.root>
       );
     };
@@ -59,22 +55,20 @@ export const DrawerV1 = compose<DrawerV1Type>({
 });
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#F5FCFF',
+  left: {
+    left: 0,
+    height: '100%',
+    width: '50%',
   },
-  content: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
+  right: {
+    right: 0,
+    height: '100%',
+    width: '50%',
   },
-  drawerContent: {
-    flex: 1,
-    alignItems: 'center',
-  },
-  text: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
+  bottom: {
+    bottom: 0,
+    borderTopRightRadius: 10,
+    borderTopLeftRadius: 10,
+    width: '100%',
   },
 });
